+++ tests/FsCodec.NewtonsoftJson.Tests/VerbatimUtf8ConverterTests.fs
@@ -67,7 +67,7 @@ module VerbatimUtf8Tests = // not a module or CI will fail for net461
         test <@ res.Contains """"d":{"embed":"\""}""" @>
         let des = JsonConvert.DeserializeObject<Batch>(res)
         let loaded = FsCodec.Core.TimelineEvent.Create(-1L, des.e[0].c, ReadOnlyMemory des.e[0].d)
-        let decoded = eventCodec.TryDecode loaded |> Option.get
+        let decoded = eventCodec.TryDecode loaded |> ValueOption.get
         input =! decoded
 
     let defaultSettings = Options.CreateDefault()
@@ -79,7 +79,7 @@ module VerbatimUtf8Tests = // not a module or CI will fail for net461
         let ser = JsonConvert.SerializeObject(e, defaultSettings)
         let des = JsonConvert.DeserializeObject<Batch>(ser, defaultSettings)
         let loaded = FsCodec.Core.TimelineEvent.Create(-1L, des.e[0].c, ReadOnlyMemory des.e[0].d)
-        let decoded = defaultEventCodec.TryDecode loaded |> Option.get
+        let decoded = defaultEventCodec.TryDecode loaded |> ValueOption.get
         x =! decoded
 
     // https://github.com/JamesNK/Newtonsoft.Json/issues/862 // doesnt apply to this case
@@ -87,7 +87,7 @@ module VerbatimUtf8Tests = // not a module or CI will fail for net461
         let x = ES { embed = "2016-03-31T07:02:00+07:00" }
         let encoded = defaultEventCodec.Encode(None,x)
         let adapted = FsCodec.Core.TimelineEvent.Create(-1L, encoded.EventType, encoded.Data, encoded.Meta, timestamp = encoded.Timestamp)
-        let decoded = defaultEventCodec.TryDecode adapted |> Option.get
+        let decoded = defaultEventCodec.TryDecode adapted |> ValueOption.get
         test <@ x = decoded @>
 
     //// NB while this aspect works, we don't support it as it gets messy when you then use the VerbatimUtf8Converter